probably will end up not using this code
after working on this, went back to the site and looked at network activity. as the task said it could be done in simple scrapy, i figured i was overcomplicating things. realized that instead of this i should probably work on figuring out the identifiers used in the api and just replicate the api requests that the site makes